Arregllo ms-inventario, arreglos varios diego

+++ MicroServicioInventario/src/main/java/com/example/MicroInventario/dto/StockInventarioDTO.java
@@ -14,16 +14,16 @@ public class StockInventarioDTO {
     public static class Request {
 
         @NotNull(message = "La tienda es obligatoria")
-        private int id_tienda;
+        private Integer id_tienda;
 
         @NotNull(message = "El producto es obligatorio")
-        private int id_producto;
+        private Long id_producto;
 
         @Min(value = 0, message = "La cantidad no puede ser negativa")
-        private int cantidad;
+        private Integer cantidad;
 
         @Min(value = 0, message = "El stock minimo no puede ser negativo")
-        private int stock_min;
+        private Integer stock_min;
 
         @NotNull(message = "El estado es obligatorio")
         private boolean estado;
@@ -33,12 +33,11 @@ public class StockInventarioDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class Response {
-
-        private int id_stock;
-        private int id_tienda;
-        private int id_producto;
-        private int cantidad;
-        private int stock_min;
+        private Integer id_stock;
+        private Integer id_tienda;
+        private Long id_producto;
+        private Integer cantidad;
+        private Integer stock_min;
         private boolean estado;
     }
 }
